requirements.txt
aiogram==2.11.2
pyotp==2.4.1
PyYAML==5.3.1
py-params==0.10.1
coloredlogs==15.0
qrcode==6.1
python-daemon==2.2.4
